fixed Zen Garden time, a vulkan synchronisation issue, restored compatibility with original save files

+++ src/todlib/TodFoley.cpp
@@ -265,7 +265,7 @@ void TodFoley::PlayFoleyPitch(FoleyType theFoleyType, float thePitch) {
         return;
 
     int aVariations = 0;
-    intptr_t aVariationsArray[10];
+    int aVariationsArray[10];
     FoleyTypeData *aFoleyData = &mFoleyTypeData[(int)theFoleyType];
     for (int i = 0; i < 10; i++) {
         if (!TestBit(aFoleyParams->mFoleyFlags, FoleyFlags::FOLEYFLAGS_DONT_REPEAT) ||
